reorganize files and separate tests in circleci
