+++ packages/api/tests/cli/test-es-cli.js
@@ -1,10 +1,10 @@
 'use strict';
 
 const test = require('ava');
-const { Search } = require('../es/search');
-const { bootstrapElasticSearch } = require('../lambdas/bootstrap');
-const es = require('../bin/es');
-const mappings = require('../models/mappings.json');
+const { Search } = require('../../es/search');
+const { bootstrapElasticSearch } = require('../../lambdas/bootstrap');
+const es = require('../../bin/es');
+const mappings = require('../../models/mappings.json');
 const get = require('lodash.get');
 
 const esIndex = 'cumulus-1';
