+++ packages/api/tests/endpoints/test-endpoints-collections.js
@@ -4,12 +4,12 @@ const sinon = require('sinon');
 const test = require('ava');
 const aws = require('@cumulus/common/aws');
 const { randomString } = require('@cumulus/common/test-utils');
-const models = require('../models');
-const bootstrap = require('../lambdas/bootstrap');
-const collectionsEndpoint = require('../endpoints/collections');
-const EsCollection = require('../es/collections');
-const { testEndpoint, fakeCollectionFactory } = require('../lib/testUtils');
-const { Search } = require('../es/search');
+const models = require('../../models');
+const bootstrap = require('../../lambdas/bootstrap');
+const collectionsEndpoint = require('../../endpoints/collections');
+const EsCollection = require('../../es/collections');
+const { testEndpoint, fakeCollectionFactory } = require('../../lib/testUtils');
+const { Search } = require('../../es/search');
 
 process.env.CollectionsTable = randomString();
 process.env.stackName = randomString();
