+++ packages/api/tests/endpoints/test-endpoints-granules.js
@@ -8,12 +8,12 @@ const { StepFunction } = require('@cumulus/ingest/aws');
 const { CMR } = require('@cumulus/cmrjs');
 const { DefaultProvider } = require('@cumulus/ingest/crypto');
 const { randomString } = require('@cumulus/common/test-utils');
-const models = require('../models');
-const bootstrap = require('../lambdas/bootstrap');
-const granuleEndpoint = require('../endpoints/granules');
-const indexer = require('../es/indexer');
-const { testEndpoint, fakeGranuleFactory } = require('../lib/testUtils');
-const { Search } = require('../es/search');
+const models = require('../../models');
+const bootstrap = require('../../lambdas/bootstrap');
+const granuleEndpoint = require('../../endpoints/granules');
+const indexer = require('../../es/indexer');
+const { testEndpoint, fakeGranuleFactory } = require('../../lib/testUtils');
+const { Search } = require('../../es/search');
 const xml2js = require('xml2js');
 const { xmlParseOptions } = require('@cumulus/cmrjs/utils');
