+++ packages/api/tests/endpoints/test-endpoints-pdrs.js
@@ -3,12 +3,12 @@
 const test = require('ava');
 const aws = require('@cumulus/common/aws');
 const { randomString } = require('@cumulus/common/test-utils');
-const models = require('../models');
-const bootstrap = require('../lambdas/bootstrap');
-const pdrEndpoint = require('../endpoints/pdrs');
-const indexer = require('../es/indexer');
-const { testEndpoint, fakePdrFactory } = require('../lib/testUtils');
-const { Search } = require('../es/search');
+const models = require('../../models');
+const bootstrap = require('../../lambdas/bootstrap');
+const pdrEndpoint = require('../../endpoints/pdrs');
+const indexer = require('../../es/indexer');
+const { testEndpoint, fakePdrFactory } = require('../../lib/testUtils');
+const { Search } = require('../../es/search');
 
 // create all the variables needed across this test
 let esClient;
