+++ packages/api/tests/endpoints/test-endpoints-providers.js
@@ -2,12 +2,12 @@
 
 const test = require('ava');
 
-const bootstrap = require('../lambdas/bootstrap');
-const models = require('../models');
+const bootstrap = require('../../lambdas/bootstrap');
+const models = require('../../models');
 const { randomString } = require('@cumulus/common/test-utils');
-const providerEndpoint = require('../endpoints/providers');
-const { testEndpoint } = require('../lib/testUtils');
-const { Search } = require('../es/search');
+const providerEndpoint = require('../../endpoints/providers');
+const { testEndpoint } = require('../../lib/testUtils');
+const { Search } = require('../../es/search');
 
 
 process.env.ProvidersTable = randomString();
