+++ packages/api/tests/lambdas/test-kinesis-consumer.js
@@ -7,12 +7,12 @@ const test = require('ava');
 const { randomString } = require('@cumulus/common/test-utils');
 const { SQS } = require('@cumulus/ingest/aws');
 const { s3, recursivelyDeleteS3Bucket } = require('@cumulus/common/aws');
-const { getKinesisRules, handler } = require('../lambdas/kinesis-consumer');
+const { getKinesisRules, handler } = require('../../lambdas/kinesis-consumer');
 
-const manager = require('../models/base');
-const Collection = require('../models/collections');
-const Rule = require('../models/rules');
-const Provider = require('../models/providers');
+const manager = require('../../models/base');
+const Collection = require('../../models/collections');
+const Rule = require('../../models/rules');
+const Provider = require('../../models/providers');
 const testCollectionName = 'test-collection';
 
 const ruleTableParams = {
